PR 타입
참가자수 증감 기능 추가

반영 브랜치
Hotfix_sortBy -> dev

변경 사항
참가자수 정렬 오류 해결위해, 참가자수 증감되도록 url 추가로 만들었습니다.

+++ src/main/java/com/hanghae/baedalfriend/dto/responsedto/PostResponseDto.java
@@ -28,7 +28,7 @@ public class PostResponseDto {
     private String deliveryTime; // 배달시간
     private Long deliveryFee; // 배달요금
     private Long maxCapacity; // 최대인원
-    private int participantNumber; // 참여자수
+    private Long participantNumber; // 참여자수
     private String nickname; // 닉네임
     private String profileURL; // 프로필사진
     private String gatherName; // 모이는 장소 이름
